+++ CHANGELOG.md
@@ -772,13 +772,24 @@
 - Benefitting countries can be imported via CSV upload
 - Tidy up Report Variance tab to only show activities that have a variance
 
-## [unreleased]
+## [release-66] - 2021-08-10
 
 - Make sure users are active even after login
 - Add breadcrumb trail for activities
 - Adjust XML exports to use new benefitting countries field
+- No longer collect geography, recipient country, recipient region, and intended beneficiaries (deprecated by collecting benefitting countries)
 
-[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-65...HEAD
+## [release-67] - 2021-08-12
+
+- Remove legacy geography steps
+- Backfill benefitting_countries where possible
+- Add horizontal margin inside table wrappers
+
+## [unreleased]
+
+[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-67...HEAD
+[release-67]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-66...release-67
+[release-66]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-65...release-66
 [release-65]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-64...release-65
 [release-64]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-63...release-64
 [release-63]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-62...release-63